pyproject.toml
[project]
name = "voice-speaker-recognition"
version = "0.1.0"
description = "Voice assistant with MLX Whisper transcription and speaker recognition"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["voice", "speech", "transcription", "speaker-recognition", "mlx", "whisper"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    # Core ML/AI
    "mlx-whisper>=0.4.2",
    "pyannote-audio>=3.3.2",
    "anthropic>=0.18.0",
    
    # Audio processing
    "pyaudio>=0.2.13",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    
    # System interaction
    "pynput>=1.8.1",
    
    # Optional memory system
    "honcho-ai>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/voice-speaker-recognition"
Documentation = "https://github.com/yourusername/voice-speaker-recognition#readme"
Repository = "https://github.com/yourusername/voice-speaker-recognition"
Issues = "https://github.com/yourusername/voice-speaker-recognition/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"